Update Go backend: rename Pandora to Prizrak and snakem982 to legiz-ru
