+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8"/>
     <link rel="icon" href="/icon.ico"/>
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <title>Pandora-Box</title>
+    <title>Prizrak-Box</title>
 </head>
 <body>
 <div id="app"></div>
