+++ src-go/api/handlers/dns.go
@@ -5,10 +5,10 @@ import (
 	"github.com/go-chi/render"
 	"github.com/metacubex/mihomo/hub/executor"
 	"github.com/metacubex/mihomo/log"
-	"github.com/snakem982/pandora-box/api/models"
-	"github.com/snakem982/pandora-box/internal"
-	"github.com/snakem982/pandora-box/pkg/cache"
-	"github.com/snakem982/pandora-box/pkg/constant"
+	"github.com/legiz-ru/prizrak-box/api/models"
+	"github.com/legiz-ru/prizrak-box/internal"
+	"github.com/legiz-ru/prizrak-box/pkg/cache"
+	"github.com/legiz-ru/prizrak-box/pkg/constant"
 	"net/http"
 )
