+++ src-go/api/handlers/rule.go
@@ -9,12 +9,12 @@ import (
 	"github.com/go-chi/render"
 	"github.com/metacubex/mihomo/hub/executor"
 	"github.com/metacubex/mihomo/log"
-	"github.com/snakem982/pandora-box/api/models"
-	"github.com/snakem982/pandora-box/internal"
-	"github.com/snakem982/pandora-box/pkg/cache"
-	"github.com/snakem982/pandora-box/pkg/constant"
-	sys "github.com/snakem982/pandora-box/pkg/sys/proxy"
-	"github.com/snakem982/pandora-box/pkg/utils"
+	"github.com/legiz-ru/prizrak-box/api/models"
+	"github.com/legiz-ru/prizrak-box/internal"
+	"github.com/legiz-ru/prizrak-box/pkg/cache"
+	"github.com/legiz-ru/prizrak-box/pkg/constant"
+	sys "github.com/legiz-ru/prizrak-box/pkg/sys/proxy"
+	"github.com/legiz-ru/prizrak-box/pkg/utils"
 )
 
 func Rule(r chi.Router) {
