+++ src-go/api/job/refresh.go
@@ -2,13 +2,13 @@ package job
 
 import (
 	"github.com/metacubex/mihomo/log"
-	"github.com/snakem982/pandora-box/api/models"
-	"github.com/snakem982/pandora-box/internal"
-	"github.com/snakem982/pandora-box/pkg/cache"
-	"github.com/snakem982/pandora-box/pkg/constant"
-	"github.com/snakem982/pandora-box/pkg/cron"
-	"github.com/snakem982/pandora-box/pkg/proxy"
-	"github.com/snakem982/pandora-box/pkg/utils"
+	"github.com/legiz-ru/prizrak-box/api/models"
+	"github.com/legiz-ru/prizrak-box/internal"
+	"github.com/legiz-ru/prizrak-box/pkg/cache"
+	"github.com/legiz-ru/prizrak-box/pkg/constant"
+	"github.com/legiz-ru/prizrak-box/pkg/cron"
+	"github.com/legiz-ru/prizrak-box/pkg/proxy"
+	"github.com/legiz-ru/prizrak-box/pkg/utils"
 	"strconv"
 	"strings"
 	"sync"
