+++ src-go/pkg/sys/proxy/proxy_linux.go
@@ -3,7 +3,7 @@ package sys
 import (
 	"bytes"
 	"fmt"
-	sys "github.com/snakem982/pandora-box/pkg/sys/cmd"
+	sys "github.com/legiz-ru/prizrak-box/pkg/sys/cmd"
 	"strconv"
 	"strings"
 )
